+++ packages/core/editor/src/YooptaEditor.tsx
@@ -1,23 +1,10 @@
-import EventEmitter from 'eventemitter3';
-import { YooptaContextProvider } from './contexts/YooptaContext/YooptaContext';
-import { getDefaultYooptaChildren } from './components/Editor/utils';
 import { Editor } from './components/Editor/Editor';
 import { CSSProperties, useMemo, useState } from 'react';
 import { SlateElement, YooEditor, YooptaBlockData, YooptaContentValue } from './editor/types';
-import { Plugin } from './plugins/types';
-import { Tools, ToolsProvider } from './contexts/YooptaContext/ToolsContext';
-import {
-  buildBlocks,
-  buildBlockShortcuts,
-  buildBlockSlateEditors,
-  buildCommands,
-  buildMarks,
-  buildPlugins,
-} from './utils/editorBuilders';
+import { Tools } from './contexts/YooptaContext/ToolsContext';
 import { YooptaPlugin } from './plugins';
 import { YooptaMark } from './marks';
-import { FakeSelectionMark } from './marks/FakeSelectionMark';
-import { generateId } from './utils/generateId';
+import { YooptaEditorProvider } from './YooptaEditorProvider';
 
 type Props = {
   id?: string;
@@ -36,30 +23,12 @@ type Props = {
   style?: CSSProperties;
 };
 
-const DEFAULT_VALUE: Record<string, YooptaBlockData> = getDefaultYooptaChildren();
-const eventEmitter = new EventEmitter();
-
-const Events = {
-  on: (event, fn) => eventEmitter.on(event, fn),
-  once: (event, fn) => eventEmitter.once(event, fn),
-  off: (event, fn) => eventEmitter.off(event, fn),
-  emit: (event, payload) => eventEmitter.emit(event, payload),
-};
-
-function validateInitialValue(value: any): boolean {
-  if (!value) return false;
-  if (typeof value !== 'object') return false;
-  if (Object.keys(value).length === 0) return false;
-
-  return true;
-}
-
 const YooptaEditor = ({
   id,
   editor,
   value,
-  marks: marksProps,
-  plugins: pluginsProps,
+  marks,
+  plugins,
   autoFocus,
   className,
   tools,
@@ -70,65 +39,28 @@ const YooptaEditor = ({
   width,
   style,
 }: Props) => {
-  const applyChanges = () => {
-    setEditorState((prev) => ({ ...prev, version: prev.version + 1 }));
-  };
-
-  const marks = useMemo(() => {
-    if (marksProps) return [FakeSelectionMark, ...marksProps];
-    return [FakeSelectionMark];
-  }, [marksProps]);
-
-  const plugins = useMemo(() => {
-    return pluginsProps.map((plugin) => plugin.getPlugin as Plugin<Record<string, SlateElement>>);
-  }, [pluginsProps]);
-
-  const [editorState, setEditorState] = useState<{ editor: YooEditor; version: number }>(() => {
-    if (!editor.id) editor.id = id || generateId();
-    editor.applyChanges = applyChanges;
-    editor.readOnly = readOnly || false;
-    if (marks) editor.formats = buildMarks(editor, marks);
-    editor.blocks = buildBlocks(editor, plugins);
-
-    const isValueValid = validateInitialValue(value);
-    if (!isValueValid && typeof value !== 'undefined') {
-      // [TODO] - add link to documentation
-      console.error(
-        `Initial value is not valid. Should be an object with blocks. You passed: ${JSON.stringify(value)}`,
-      );
-    }
-
-    editor.children = (isValueValid ? value : DEFAULT_VALUE) as YooptaContentValue;
-    editor.blockEditorsMap = buildBlockSlateEditors(editor);
-    editor.shortcuts = buildBlockShortcuts(editor);
-    editor.plugins = buildPlugins(plugins);
-    editor.commands = buildCommands(editor, plugins);
-
-    editor.on = Events.on;
-    editor.once = Events.once;
-    editor.off = Events.off;
-    editor.emit = Events.emit;
-
-    return { editor, version: 0 };
-  });
-
   return (
-    <YooptaContextProvider editorState={editorState}>
-      <ToolsProvider tools={tools}>
-        <Editor
-          placeholder={placeholder}
-          marks={marks}
-          autoFocus={autoFocus}
-          className={className}
-          selectionBoxRoot={selectionBoxRoot}
-          width={width}
-          style={style}
-        >
-          {children}
-        </Editor>
-      </ToolsProvider>
-    </YooptaContextProvider>
+    <YooptaEditorProvider
+      id={id}
+      editor={editor}
+      value={value}
+      marks={marks}
+      plugins={plugins}
+      tools={tools}
+      readOnly={readOnly}
+    >
+      <Editor
+        placeholder={placeholder}
+        marks={marks}
+        autoFocus={autoFocus}
+        className={className}
+        selectionBoxRoot={selectionBoxRoot}
+        width={width}
+        style={style}
+      >
+        {children}
+      </Editor>
+    </YooptaEditorProvider>
   );
 };
-
 export { YooptaEditor };
